+++ backend/rs/ic_vetkeys/tests/utils.rs
@@ -1,8 +1,8 @@
+use hex_literal::hex;
 use ic_bls12_381::*;
 use ic_vetkeys::*;
 use ic_vetkeys_test_utils::*;
 use rand::Rng;
-use hex_literal::hex;
 
 #[test]
 fn test_hkdf_test_vector() {
@@ -28,7 +28,10 @@ fn test_hkdf_test_vector() {
 
 #[test]
 fn test_is_valid_transport_public_key() {
-    assert_eq!(is_valid_transport_public_key_encoding(&hex::decode("F00F00F00F00").unwrap()), false);
+    assert_eq!(
+        is_valid_transport_public_key_encoding(&hex::decode("F00F00F00F00").unwrap()),
+        false
+    );
     assert_eq!(is_valid_transport_public_key_encoding(&hex::decode("97f1d3a73197d7942695638c4fa9ac0fc3688c4f9774b905a14e3a3f171bac586c55e83ff97a1aeffb3af00adb22c6bb").unwrap()), true);
     assert_eq!(is_valid_transport_public_key_encoding(&hex::decode("c00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000").unwrap()), true);
 }
@@ -68,7 +71,8 @@ fn test_bls_signature_verification() {
 fn test_bls_signature_verification_using_identity() {
     // Check that the identity element is rejected as a public key
 
-    let dpk = DerivedPublicKey::deserialize(&ic_bls12_381::G2Affine::identity().to_compressed()).unwrap();
+    let dpk =
+        DerivedPublicKey::deserialize(&ic_bls12_381::G2Affine::identity().to_compressed()).unwrap();
 
     let msg = b"wrong message";